Move bip conf to bip folder
