+++ bip_utils/addr/fil_addr.py
@@ -23,7 +23,7 @@
 from typing import Any, Union
 from bip_utils.addr.iaddr_encoder import IAddrEncoder
 from bip_utils.addr.utils import AddrUtils
-from bip_utils.conf import Bip44Filecoin
+from bip_utils.bip.conf import Bip44Filecoin
 from bip_utils.ecc import IPublicKey
 from bip_utils.utils import Base32Encoder, ConvUtils, CryptoUtils
