+++ bip_utils/bip/bip32/bip32_ed25519_slip.py
@@ -25,7 +25,7 @@ from bip_utils.bip.bip32.bip32_base import Bip32Base
 from bip_utils.bip.bip32.bip32_ed25519_slip_base import Bip32Ed25519SlipBaseConst, Bip32Ed25519SlipBase
 from bip_utils.bip.bip32.bip32_key_data import Bip32KeyIndex
 from bip_utils.bip.bip32.bip32_path import Bip32Path
-from bip_utils.conf import Bip44BitcoinMainNet, KeyNetVersions
+from bip_utils.bip.conf import Bip44BitcoinMainNet, KeyNetVersions
 from bip_utils.ecc import EllipticCurveTypes, IPrivateKey
