+++ bip_utils/bip/bip32/bip32_key_data.py
@@ -23,7 +23,7 @@
 from __future__ import annotations
 from typing import Union
 from bip_utils.bip.bip32.bip32_utils import Bip32Utils
-from bip_utils.conf import KeyNetVersions
+from bip_utils.bip.conf import KeyNetVersions
 from bip_utils.utils import ConvUtils
