+++ bip_utils/bip/bip32/bip32_key_ser.py
@@ -24,7 +24,7 @@ from typing import Tuple
 from bip_utils.base58 import Base58Decoder, Base58Encoder
 from bip_utils.bip.bip32.bip32_ex import Bip32KeyError
 from bip_utils.bip.bip32.bip32_key_data import Bip32FingerPrint, Bip32Depth, Bip32KeyIndex, Bip32KeyData
-from bip_utils.conf import KeyNetVersions
+from bip_utils.bip.conf import KeyNetVersions
 from bip_utils.ecc import IPublicKey, IPrivateKey
 from bip_utils.utils import ConvUtils
