+++ bip_utils/bip/bip32/bip32_nist256p1.py
@@ -25,7 +25,7 @@ from bip_utils.bip.bip32.bip32_base import Bip32Base
 from bip_utils.bip.bip32.bip32_ecdsa_base import Bip32EcdsaBase
 from bip_utils.bip.bip32.bip32_key_data import Bip32KeyIndex
 from bip_utils.bip.bip32.bip32_path import Bip32Path
-from bip_utils.conf import Bip44BitcoinMainNet, KeyNetVersions
+from bip_utils.bip.conf import Bip44BitcoinMainNet, KeyNetVersions
 from bip_utils.ecc import EllipticCurveTypes, IPrivateKey
