+++ bip_utils/bip/conf/bip49/__init__.py
@@ -1,0 +1,2 @@
+from bip_utils.bip.conf.bip49.bip49_conf import *
+from bip_utils.bip.conf.bip49.bip49_conf_getter import Bip49ConfGetter
