+++ bip_utils/bip/conf/bip84/bip84_conf_getter.py
@@ -21,9 +21,9 @@
 
 # Imports
 from typing import Dict
-from bip_utils.conf.common import BipCoinConf
-from bip_utils.conf.bip44.bip44_coins import Bip44Coins
-from bip_utils.conf.bip84.bip84_conf import *
+from bip_utils.bip.conf.common import BipCoinConf
+from bip_utils.bip.conf.bip44.bip44_coins import Bip44Coins
+from bip_utils.bip.conf.bip84.bip84_conf import *
 
 
 class Bip84ConfGetterConst:
