+++ bip_utils/bip/conf/common/__init__.py
@@ -1,0 +1,7 @@
+from bip_utils.bip.conf.common.addr_types import AddrTypes
+from bip_utils.bip.conf.common.bip32_types import Bip32Types
+from bip_utils.bip.conf.common.bip_coin_conf import BipCoinConf
+from bip_utils.bip.conf.common.bip_bitcoin_cash_conf import BipBitcoinCashConf
+from bip_utils.bip.conf.common.bip_litecoin_conf import BipLitecoinConf
+from bip_utils.bip.conf.common.bip_conf_const import *
+from bip_utils.bip.conf.common.key_net_versions import KeyNetVersions
