+++ bip_utils/bip/conf/common/bip_coin_conf.py
@@ -21,9 +21,9 @@
 
 # Imports
 from typing import Any, Dict, Optional, Union
-from bip_utils.conf.common.addr_types import AddrTypes
-from bip_utils.conf.common.bip32_types import Bip32Types
-from bip_utils.conf.common.key_net_versions import KeyNetVersions
+from bip_utils.bip.conf.common.addr_types import AddrTypes
+from bip_utils.bip.conf.common.bip32_types import Bip32Types
+from bip_utils.bip.conf.common.key_net_versions import KeyNetVersions
 from bip_utils.utils import CoinNames
