--- bip_utils/conf/__init__.py
@@ -1,4 +1,0 @@
-from bip_utils.conf.common import *
-from bip_utils.conf.bip44 import *
-from bip_utils.conf.bip49 import *
-from bip_utils.conf.bip84 import *
